feat: Dependabot checks for GitHub Actions updates
The schedule is exactly the same as for the Node.js checks.
Each Monday trough Friday, Dependabot will check for updates.
It will check for updates at 8 o'clock in the morning, Dutch time.

You won't need to add a new label to your repository,
Dependabot will manage this for you.

For the official GitHub documentation read:
https://docs.github.com/en/free-pro-team@latest/github/administering-a-repository/keeping-your-actions-up-to-date-with-github-dependabot